+++ project/fun_fact.py
@@ -1,0 +1,8 @@
+import requests
+
+
+def fun_fact():
+    """Displays a random fun fact"""
+    responses = requests.get("https://api.aakhilv.me/fun/facts")
+    responses_json = responses.json()
+    return responses_json[0]
