+++ project/models.py
@@ -1,3 +1,4 @@
+import flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin
