+++ project/templates/settings.html
@@ -3,6 +3,6 @@
 
 <h1> Settings Page </h1>
 
-<a href="{{ url_for('home') }}"> Home</a>
+<a href="{{ url_for('main.home') }}"> Home</a>
 
 </html>
