--- templates/entries.html
@@ -1,29 +1,0 @@
-<html>
-    <head>
-
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
-
-    </head>
-    <body>
-        <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
-        {% with messages = get_flashed_messages() %}
-            {% if messages %}
-                {% for message in messages %}
-                    <b>{{message}}</b>
-                {% endfor %}
-            {% endif %}
-        {% endwith %}
-        <h1>Here are you're previous entries</h1>
-
-
-        {% for i in range(0,length) : %}
-            <div class="container p-2 m-5 bg-dark text-white">
-                <form action="/delete_entry" method='POST'>
-                    <button type='submit' name="Delete" value={{i}}> X </button>
-                </form>
-                <h3>{{user_entries[i]['title']}}</h3>
-                <p>{{user_entries[i]['post']}}</p>
-            </div>
-        {% endfor %}
-    </body>
-</html>
