--- templates/home.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-
-<head>
-    <title>Home</title>
-</head>
-
-<body>
-    <div id="nav">
-        <a href="{{ url_for('home') }}"> Home</a>
-        <a href="{{ url_for('users_entries') }}"> Entries</a>
-        <a href="{{ url_for('settings') }}"> Settings</a>
-        <a href="{{ url_for('signout') }}"> Log Out</a>
-    </div>
-    <h1>Hello, {{user}}</h1>
-
-    <div id="weather_info">
-        <p>{{ weather_info['city'] }}, {{ weather_info['country'] }}</p>
-        <p> {{ weather_info['weather'] }}</p>
-        <p>{{ weather_info['fahrenheit'] }}</p>
-    </div>
-
-    <div id="maincontent">
-        <form action="/addentry" method="POST">
-            <input type="text" placeholder="Title" name="title">
-            <input type="text" placeholder="How are you feeling today?" name="entry">
-            <input type="submit" value="Post">
-        </form>
-        <!-- next add task form -->
-    </div>
-</body>
-
-</html>
